# cmd/api/docs/swagger.yaml
basePath: /api
definitions:
...
  handler.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  ssl.AcmeClientResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      server_url:
        type: string
    type: object
  ssl.CreateAcmeClientRequest:
    properties:
      email:
        type: string
      key_type:
        type: string
      server_url:
        type: string
    required:
    - email
    - key_type
    - server_url
    type: object
  ssl.CreateDnsAccountRequest:
    properties:
      key:
        type: string
      provider_id:
        type: integer
      value:
        type: string
    required:
    - key
    - provider_id
    - value
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@etapanel.com
    name: EtaPanel Support
    url: https://github.com/EtaPanel-dev/EtaPanel
  description: EtaPanel 服务器管理面板后端API文档，提供系统监控、文件管理、Nginx配置、SSL证书管理等功能
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: EtaPanel API
  version: "1.0"
paths:
  /auth/acme/clients:
    get:
      consumes:
      - application/json
      description: 获取所有ACME客户端配置
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/ssl.AcmeClientResponse'
                  type: array
              type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: 获取ACME客户端列表
      tags:
      - ACME客户端管理
    post:
      consumes:
      - application/json
      description: 创建新的ACME客户端配置
      parameters:
      - description: ACME客户端信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ssl.CreateAcmeClientRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: 创建ACME客户端
      tags:
      - ACME客户端管理
  /auth/acme/clients/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定ID的ACME客户端配置
      parameters:
      - description: 客户端ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: 客户端不存在
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: 删除ACME客户端
      tags:
      - ACME客户端管理
    put:
      consumes:
      - application/json
      description: 更新指定ID的ACME客户端配置
      parameters:
      - description: 客户端ID
        in: path
        name: id
        required: true
        type: integer
      - description: ACME客户端信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ssl.CreateAcmeClientRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: 客户端不存在
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: 更新ACME客户端
      tags:
      - ACME客户端管理
  /auth/acme/dns:
    get:
      consumes:
      - application/json
      description: 获取所有DNS服务商账号配置
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  items:
                    type: object
                  type: array
              type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: 获取DNS账号列表
      tags:
      - DNS账号管理
    post:
      consumes:
      - application/json
      description: 创建新的DNS服务商账号配置
      parameters:
      - description: DNS账号信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ssl.CreateDnsAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: 创建DNS账号
      tags:
      - DNS账号管理
  /auth/acme/dns/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定ID的DNS服务商账号配置
      parameters:
      - description: 账号ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: 账号不存在
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: 删除DNS账号
      tags:
      - DNS账号管理
    put:
      consumes:
      - application/json
      description: 更新指定ID的DNS服务商账号配置
      parameters:
      - description: 账号ID
        in: path
        name: id
        required: true
        type: integer
      - description: DNS账号信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ssl.CreateDnsAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: 账号不存在
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: 更新DNS账号
      tags:
      - DNS账号管理
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 输入Bearer token，格式：Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...